Create Cog edit group with moderation commands that include the editing of server name, region, text channel, and voice channel, and role. Create a error handler for inputing wrong region.
These methods were made to allow the moderator to easily make moderation
changes throughout their discord server without haveing to manually do
them. The error handler was created because if the wrong region is
inputted the user should be aware that this is the case so they correct
it in the future.