Updated newadmin role by assigning the role even if no one has it previously
Added this feature becasue the method would crash if no one else has the
head admin role previously. This is because the list of members with
this role would be empty and when accessing the empty list it would
throw in error. Now if it is empty it'll just automatically add the role
instead of trying to remove it from previous members