Create clear task to run every 2 hours
This clear loop task will run every two hours and automatically clear
the queue and delete the files in the server. This was created so that
the files don't take up too much space in the back end and don't clog up
the storage. This was also created because if someone kicks the bot out
fo a channel and doesn't run stop command it will still delete the files eventually.