Fix extra mini game methods by adding missing parameter 'self'
both of the smaller mini games rock paper scissors and coin toss were
not working. They were missing the self keyword in their parameter
which is required in python classes. Without this the object instance
was not being sent out to the method causing it to fail, this should now
be fixed